Move update_celery_state to more global place
This function is called when any celery task is published, not just
tasks in casexml/apps/phone/tasks.py
