+++ corehq/celery_monitoring/signals.py
@@ -4,7 +4,10 @@ import logging
 
 from django.core.cache import cache
 
-from celery.signals import before_task_publish, task_postrun, task_prerun
+from celery.signals import after_task_publish, before_task_publish, task_postrun, task_prerun
+from celery import current_app
+
+from casexml.apps.phone.tasks import ASYNC_RESTORE_SENT
 
 from dimagi.utils.parsing import string_to_utc_datetime
 
@@ -20,6 +23,28 @@ def celery_add_time_sent(headers=None, body=None, **kwargs):
     if eta:
         eta = TimeToStartTimer.parse_iso8601(eta)
     TimeToStartTimer(task_id).start_timing(eta)
+
+
+@after_task_publish.connect
+def update_celery_state(sender=None, headers=None, **kwargs):
+    """Updates the celery task progress to "SENT"
+
+    When fetching an task from celery using the form AsyncResponse(task_id), if
+    task_id never exists, celery will not throw an error, it will just hang.
+    This function updates each task sent to celery to have a progress value of "SENT",
+    which means we can check that a task exists with the following pattern:
+    `AsyncResponse(task_id).status == "SENT"`
+
+    See
+    http://stackoverflow.com/questions/9824172/find-out-whether-celery-task-exists/10089358
+    for more info
+
+    """
+
+    task = current_app.tasks.get(sender)
+    backend = task.backend if task else current_app.backend
+
+    backend.store_result(headers['id'], None, ASYNC_RESTORE_SENT)
 
 
 @task_prerun.connect
