+++ corehq/ex-submodules/casexml/apps/phone/tasks.py
@@ -5,9 +5,8 @@ from django.conf import settings
 from django.db import connections, router
 from django.db.models import Min
 
-from celery import current_app, current_task
+from celery import current_task
 from celery.schedules import crontab
-from celery.signals import after_task_publish
 
 from casexml.apps.phone.models import SyncLogSQL
 from dimagi.utils.logging import notify_exception
@@ -44,28 +43,6 @@ def get_async_restore_payload(restore_config, domain=None, username=None):
     restore_config.async_restore_task_id_cache.invalidate()
 
     return response.name
-
-
-@after_task_publish.connect
-def update_celery_state(sender=None, headers=None, **kwargs):
-    """Updates the celery task progress to "SENT"
-
-    When fetching an task from celery using the form AsyncResponse(task_id), if
-    task_id never exists, celery will not throw an error, it will just hang.
-    This function updates each task sent to celery to have a progress value of "SENT",
-    which means we can check that a task exists with the following pattern:
-    `AsyncResponse(task_id).status == "SENT"`
-
-    See
-    http://stackoverflow.com/questions/9824172/find-out-whether-celery-task-exists/10089358
-    for more info
-
-    """
-
-    task = current_app.tasks.get(sender)
-    backend = task.backend if task else current_app.backend
-
-    backend.store_result(headers['id'], None, ASYNC_RESTORE_SENT)
 
 
 @periodic_task(
